% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_pmtiles_metadata}
\alias{rust_pmtiles_metadata}
\title{Read PMTiles header and metadata as a JSON string}
\usage{
rust_pmtiles_metadata(path)
}
\arguments{
\item{path}{Path to the .pmtiles file}
}
\description{
Read PMTiles header and metadata as a JSON string
}
\keyword{internal}
