% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freestiler-package.R
\docType{package}
\name{freestiler-package}
\alias{freestiler}
\alias{freestiler-package}
\title{freestiler: Create Vector Tiles from Spatial Data}
\description{
Create vector tile archives in PMTiles format from sf spatial data frames. Supports Mapbox Vector Tile (MVT) and MapLibre Tile (MLT) output formats. Uses a Rust backend via 'extendr' for fast, in-memory tiling with zero external system dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://walker-data.com/freestiler/}
  \item \url{https://github.com/walkerke/freestiler}
  \item Report bugs at \url{https://github.com/walkerke/freestiler/issues}
}

}
\author{
\strong{Maintainer}: Kyle Walker \email{kyle@walker-data.com}

}
\keyword{internal}
